removed old imports, unnecessary comments, debug console.logs. Replaced old logo in all locations with new logo. Moved game.model.ts interface to _models
