+++ src/app/game-review/game-review.component.html
@@ -29,7 +29,6 @@
                         <!-- [featureImg]="game.featureDescriptions[index] will show each feature image from the array of the game object because of our ngFor directive -->
                         <app-media-feature [showImage]="true"[showTitle]="false" [featureImg]="game.featureImages[index]"></app-media-feature>
                     </div>
-                    <!-- <img class="feature-img"src="../../assets/Warcraft-III-generic-image-half-size.png" alt="img" height="200px"> -->
                     <p>{{ game.featureDescriptions[index] }}</p>
                 </div>
             <hr>
