+++ src/app/game-review/game-review.component.ts
@@ -1,7 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import { ActivatedRoute } from '@angular/router';
 import { GamesService } from '../_services/games.service';
-import { Game } from '../models/game.model';
+
 @Component({
   selector: 'app-game-review',
   templateUrl: './game-review.component.html',
