+++ src/app/game-reviews-list/game-reviews-list.component.ts
@@ -1,6 +1,4 @@
 import { Component, OnInit } from '@angular/core';
-import { GamesService } from '../_services/games.service';
-import { Game } from '../models/game.model';
 
 @Component({
   selector: 'app-game-reviews-list',
@@ -9,7 +7,7 @@ import { Game } from '../models/game.model';
 })
 export class GameReviewsListComponent implements OnInit {
 
-  constructor(private gameService: GamesService) { }
+  constructor() { }
 
   ngOnInit(): void {
   }
