+++ src/app/game-reviews-list/modify-game-list/modify-game-list.component.html
@@ -76,22 +76,3 @@
     </div>
   </div>
   
-
-
-<!-- 
-  <div formArrayName="featureImages">
-    <div *ngFor="let featureI of featureImages.controls; let index = index">
-      <label>Image: <input type="text" [formControlName]="index" ></label>
-      <hr>
-    </div>
-  </div>
-  <div formArrayName="featureDescriptions">
-    <div *ngFor="let featureD of featureDescriptions.controls; let index = index">
-      <label>Description:<textarea class="description-area" type="text" rows="3" cols="25"
-      [formControlName]="index" Description"></textarea></label>
-      <hr>
-    </div>
-  </div> -->
-
-
-  <!-- <label for="featureDescriptions-{{ index }}">Feature Description:</label> -->
