+++ src/app/login/login.component.html
@@ -17,7 +17,7 @@
           </div>
           <div class="card-body" style="background-color: #333;">
                 <div class="card-title" style="margin: 20px">
-                    <div><img class="logo" src="../../assets/WARCRAFT_3-logo-rounded-2.png"></div>
+                    <div><img class="logo" src="../../assets/WARCRAFT 3-logo-rounded-v3.png"></div>
                     <div style="padding: 5px;"><img class="websiteName" src="../../assets/warcraftlegends.png"></div>
                 </div>
                 <form>
