+++ src/app/media-feature/media-feature.component.css
@@ -18,7 +18,6 @@
         z-index: 1;
         width: 100%;
         top: 20px;
-        /*left: 50px; */
         font-size: 2.5em;
 }
 .previewImg{
@@ -26,9 +25,6 @@
         margin: 0px;
         width: 100%;
 }
-
-
-
 /*applys the box shadow on top of the video element*/
 .shadow-effect{
         position: absolute;
