+++ src/app/navbar/navbar.component.html
@@ -4,13 +4,13 @@
         <!------------------------------- Left Side Of Navbar ---------------------------------->
         <div class="navbar-header">
           <a class="navbar-brand" href="">
-            <img class="logo" alt="Brand" src="../../assets/WARCRAFT_3-logo-rounded-2.png"><img src="../../assets/warcraftlegends.png" style="width: 200px; margin-left: 10px;">
+            <img class="logo" alt="Brand" src="../../assets/WARCRAFT 3-logo-rounded-v3.png"><img src="../../assets/warcraftlegends.png" style="width: 200px; margin-left: 10px;">
           </a>
         </div>
         <!----------------------------------- Middle Of Navbar ---------------------------------->
         <ul class="navbar-nav me-auto">
           <li class="nav-item">
-            <a class="nav-link active" routerLink="" routerLinkActive="active"(click)="printHTTP()">Home
+            <a class="nav-link active" routerLink="" routerLinkActive="active">Home
               <span class="visually-hidden">(current)</span>
             </a>
           </li>
